requirements.txt
asgiref==3.3.1
Django==3.1.7
django-admin-sortable2==0.7.8
django-tinymce==3.2.0
Pillow==7.2.0
pytz==2021.1
sqlparse==0.4.1
